restructured project structure, used src/ type of directories
