+++ .gitignore
@@ -3,14 +3,14 @@ $cat gitignore
 # ignore all the informational hidden folders
 /.git
 /.idea
-qteasy/__pycache__/core.cpython-36.pyc
+src/__pycache__/core.cpython-36.pyc
 *.pyc
 .idea/workspace.xml
 *.xml
 *.ipynb
 .ipynb_checkpoints/
 *.pyc
-qteasy/__pycache__/
+src/__pycache__/
 .idea/
 **.csv
 **.pyc
@@ -22,24 +22,24 @@ examples/strategy_optimization.ipynb
 */qteasy/data/
 **/data/
 *.dat
-qteasy/data/close.dat
-qteasy/.DS_Store
+src/data/close.dat
+src/.DS_Store
 .DS_Store
 ani_plot.py
 .idea/misc.xml
 .idea/modules.xml
 .idea/qteasy.iml
 .idea/vcs.xml
-qteasy/qteasy.cnf
-qteasy/qteasy.log
-qteasy/log/qteasy.log
-qteasy/log/*
+src/qteasy.cnf
+src/qteasy.log
+src/log/qteasy.log
+src/log/*
 *.log
-**qteasy/log/
+src/log/
 *.cnf
-qteasy/config/qt_config.cfg
-qteasy/config/saved.cfg
-qteasy/config/saved2.cfg
+src/config/qt_config.cfg
+src/config/saved.cfg
+src/config/saved2.cfg
 qteasy.cfg
 config/
 syslog/
