+++ setup.cfg
@@ -11,10 +11,10 @@
 
 [metadata]
 name = qteasy
-version = 1.1.4
+version = 1.1.6
 author = Jackie PENG
 author_email = jackie.pengzhao@gmail.com
-Maintainer = Jackie PENG
+maintainer = Jackie PENG
 description = A fast quantitative investment tool kit
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -42,8 +42,7 @@ classifiers =
 zip_safe = False
 include_package_data = True
 packages = qteasy
-;package_dir =
-;    = qteasy
+
 python_requires = >=3.6, <3.9
 install_requires =
     pandas>=1.4.0
@@ -58,6 +57,3 @@ install_requires =
 database = pymysql >= 1.0.0
 hdf = pytables >= 3.6.1
 feather = pyarrow >= 3
-
-[options.packages.find]
-where = qteasy
