+++ src/qteasy/core.py
@@ -17,13 +17,13 @@ import datetime
 
 import qteasy
 from .history import get_history_panel, HistoryPanel
-from .utilfuncs import str_to_list, regulate_date_format, match_ts_code
-from .utilfuncs import next_market_trade_day
-from .utilfuncs import AVAILABLE_ASSET_TYPES, _partial_lev_ratio
+from utilfuncs import str_to_list, regulate_date_format, match_ts_code
+from utilfuncs import next_market_trade_day
+from utilfuncs import AVAILABLE_ASSET_TYPES, _partial_lev_ratio
 from .finance import CashPlan
 from .qt_operator import Operator
-from .visual import _plot_loop_result, _loop_report_str, _print_test_result
-from .visual import _plot_test_result
+from visual import _plot_loop_result, _loop_report_str, _print_test_result
+from visual import _plot_test_result
 from ._arg_validators import _update_config_kwargs, ConfigDict
 from ._arg_validators import QT_CONFIG, _vkwargs_to_text
 from .optimization import _evaluate_all_parameters, _evaluate_one_parameter
@@ -1326,7 +1326,7 @@ def save_config(config=None, file_name=None, overwrite=True, initial_config=Fals
 
     file_name = _check_config_file_name(file_name=file_name, allow_default_name=initial_config)
 
-    config_path = os.path.join(QT_ROOT_PATH, 'config/')
+    config_path = os.path.join(QT_ROOT_PATH, '../config/')
     if not os.path.exists(config_path):
         os.makedirs(config_path, exist_ok=False)
     if overwrite:
@@ -1367,7 +1367,7 @@ def load_config(config=None, file_name=None):
 
     file_name = _check_config_file_name(file_name=file_name, allow_default_name=False)
 
-    config_path = os.path.join(QT_ROOT_PATH, 'config/')
+    config_path = os.path.join(QT_ROOT_PATH, '../config/')
     try:
         with open(os.path.join(config_path, file_name), 'rb') as f:
             saved_config = pickle.load(f)
