+++ src/qteasy/space.py
