+++ src/qteasy/strategy.py
