+++ src/qteasy/tafuncs.py
