+++ src/qteasy/tsfuncs.py
