+++ src/qteasy/utilfuncs.py
